requirements.txt
Django~=4.0.3
djangorestframework==3.13.1
drf-yasg==1.20.0
psycopg2-binary~=2.9.3
django-filter==21.1
